New new build system using autoconf + GNU make
New build system using autoconf and GNU make (and in the future: libtool).

This is partially based on the ScummVM build system (which in turn borrows
from the pentagram build system), and also in parts on the git.git build
system.

This is less complex than automake; while it requires a little bit more work
from us for a few things (e.g. dependency tracking), it is much simpler to
understand, and hence to debug, as it doesn't try to hide the complexity
behind m4 macro tricks.

Among the key benefits, it enables us to handle HPC-GAP mode in clever ways...